+++ src/Client/State/Window.hs
@@ -19,6 +19,7 @@ module Client.State.Window
   , winUnread
   , winTotal
   , winMention
+  , winMarker
 
   -- * Window lines
   , WindowLine(..)
@@ -36,6 +37,8 @@ module Client.State.Window
   , emptyWindow
   , addToWindow
   , windowSeen
+  , windowActivate
+  , windowDeactivate
   ) where
 
 import           Client.Image.PackedImage
@@ -62,18 +65,22 @@ data WindowLines
 -- | A 'Window' tracks all of the messages and metadata for a particular
 -- message buffer.
 data Window = Window
-  { _winMessages :: !WindowLines  -- ^ Messages to display, newest first
-  , _winUnread   :: !Int          -- ^ Messages added since buffer was visible
-  , _winTotal    :: !Int          -- ^ Messages in buffer
-  , _winMention  :: !Bool         -- ^ Indicates an important event is unread
+  { _winMessages :: !WindowLines   -- ^ Messages to display, newest first
+  , _winMarker   :: !(Maybe Int)   -- ^ Location of line drawn to indicate newer messages
+  , _winUnread   :: !Int           -- ^ Messages added since buffer was visible
+  , _winTotal    :: !Int           -- ^ Messages in buffer
+  , _winMention  :: !WindowLineImportance -- ^ Indicates an important event is unread
   }
+
+data ActivityLevel = NoActivity | NormalActivity | HighActivity
+  deriving (Eq, Ord, Read, Show)
 
 -- | Flag for the important of a message being added to a window
 data WindowLineImportance
   = WLBoring -- ^ Don't update unread count
   | WLNormal -- ^ Increment unread count
   | WLImportant -- ^ Increment unread count and set important flag
-  deriving (Eq, Show)
+  deriving (Eq, Ord, Show, Read)
 
 makeLenses ''Window
 makeLenses ''WindowLine
@@ -94,9 +101,10 @@ wlFullImage = wlFullImage' . _Image'
 emptyWindow :: Window
 emptyWindow = Window
   { _winMessages = Nil
+  , _winMarker   = Nothing
   , _winUnread   = 0
   , _winTotal    = 0
-  , _winMention  = False
+  , _winMention  = WLBoring
   }
 
 -- | Adds a given line to a window as the newest message. Window's
@@ -105,16 +113,31 @@ addToWindow :: WindowLine -> Window -> Window
 addToWindow !msg !win = Window
     { _winMessages = msg :- view winMessages win
     , _winTotal    = view winTotal win + 1
-    , _winUnread   = view winUnread win
-                   + (if view wlImportance msg == WLBoring then 0 else 1)
-    , _winMention  = view winMention win
-                  || view wlImportance msg == WLImportant
+    , _winMarker   = do i <- view winMarker win; return $! i+1
+    , _winUnread   = if view wlImportance msg == WLBoring
+                     then view winUnread win
+                     else view winUnread win + 1
+    , _winMention  = max (view winMention win) (view wlImportance msg)
     }
 
 -- | Update the window clearing the unread count and important flag.
 windowSeen :: Window -> Window
 windowSeen = set winUnread 0
-           . set winMention False
+           . set winMention WLBoring
+
+
+-- | Update the window when it first becomes active. If only /boring/
+-- messages have been added since last time the marker will be hidden.
+windowActivate :: Window -> Window
+windowActivate win
+  | view winUnread win == 0 = set winMarker Nothing win
+  | otherwise               = win
+
+
+-- | Update the window when it becomes inactive. This resets the activity
+-- marker to the bottom of the window.
+windowDeactivate :: Window -> Window
+windowDeactivate = set winMarker (Just 0)
 
 
 instance Each WindowLines WindowLines WindowLine WindowLine where